Fix build for macOS GPUI
GPUI requires metal, which is not shipped by the Xcode Command Line Utilities,
only shipped by Xcode. Even then Xcode 26 doesnt ship it with Xcode, you need the
component, which is too much of an L for Nix-darwin users.

The fix is to add a feature to GPUI for macOS where is gets the metal shaders,
or something I honestly don't really understand how this fixes it.